langgraph/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Off-Grid AI Agent - LangGraph router and tools"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=1.2.7",
    "langchain-core>=1.2.7",
    "langchain-ollama>=1.0.1",
    "langchain-community>=0.3.0",
    "langchain-text-splitters>=0.3.0",
    "pypdf>=5.0.0",
    "faiss-cpu>=1.9.0",
    "langgraph>=1.0.7",
    "ollama>=0.6.1",
    "fastapi>=0.128.0",
    "uvicorn>=0.40.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "twilio>=9.10.0",
    "geopy>=2.4.0",
    "httpx>=0.27.0",
    "langfuse>=3.0.0",
    "webapp",
    "smolagents>=1.24.0",
    "bs4>=0.0.2",
    "ddgs>=9.10.0",
    "langchain-anthropic>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-google-genai>=2.0.0",
]

[tool.uv.sources]
webapp = { workspace = true }
